backend/pyproject.toml
[project]
name = "career-evaluation-backend"
version = "0.1.0"
description = "FastAPI backend for career evaluation (ports Next.js /api/*)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
    "google-genai>=1.0",
    "supabase>=2.9",
    "PyJWT[crypto]>=2.9",
    "openpyxl>=3.1",
    "sse-starlette>=2.1",
    "python-dateutil>=2.9",
    "orjson>=3.10",
    "httpx>=0.27",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "freezegun>=1.5",
    "ruff>=0.7",
    "mypy>=1.13",
    "anyio>=4.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
